+++ src/Crypto/Argon2.hs
@@ -1,14 +1,15 @@
 {-# LANGUAGE DeriveDataTypeable #-}
-{-# LANGUAGE DeriveGeneric #-}
-{-# LANGUAGE ForeignFunctionInterface #-}
-{-# LANGUAGE RecordWildCards #-}
+{-# LANGUAGE DeriveGeneric      #-}
+{-# LANGUAGE OverloadedStrings  #-}
+{-# LANGUAGE RecordWildCards    #-}
+{-# LANGUAGE Trustworthy        #-}
 
 {-|
 
 "Crypto.Argon2" provides bindings to the
 <https://github.com/P-H-C/phc-winner-argon2 reference implementation> of Argon2,
 the password-hashing function that won the
-<https://password-hashing.net/ Password Hashing Competition (PHC)>. 
+<https://password-hashing.net/ Password Hashing Competition (PHC)>.
 
 The main entry points to this module are 'hashEncoded', which produces a
 crypt-like ASCII output; and 'hash' which produces a 'BS.ByteString' (a stream
@@ -21,27 +22,30 @@ For access directly to the C interface, see "Crypto.Argon2.FFI".
 
 -}
 module Crypto.Argon2
-       ( -- * Computing hashes
-         hashEncoded, hash,
-         -- * Verification
-         verify,
-         -- * Configuring hashing
-         HashOptions(..), Argon2Variant(..), defaultHashOptions,
-         -- * Exceptions
-         Argon2Exception(..))
-       where
+    ( -- * Computing hashes
+      hashEncoded
+    , hash
+      -- * Verification
+    , verifyEncoded
+      -- * Configuring hashing
+    , HashOptions(..)
+    , Argon2Variant(..)
+    , Argon2Version(..)
+    , defaultHashOptions
+      -- * Exceptions
+    , Argon2Exception(..)
+    ) where
 
-import GHC.Generics (Generic)
-import Control.Exception
-import Data.Typeable
-import Foreign
-import Foreign.C
-import Numeric.Natural
-import System.IO.Unsafe (unsafePerformIO)
+import           Control.Exception
 import qualified Crypto.Argon2.FFI as FFI
-import qualified Data.ByteString as BS
-import qualified Data.Text as T
-import qualified Data.Text.Encoding as T
+import qualified Data.ByteString   as BS
+import qualified Data.Text.Short   as TS
+import           Data.Typeable
+import           Foreign
+import           Foreign.C
+import           GHC.Generics      (Generic)
+import           Numeric.Natural
+import           System.IO.Unsafe  (unsafePerformIO)
 
 -- | Which variant of Argon2 to use. You should choose the variant that is most
 -- applicable to your intention to hash inputs.
@@ -50,24 +54,38 @@ data Argon2Variant
              -- for password hashing and password-based key derivation. Argon2i
              -- is slower as it makes more passes over the memory to protect from
              -- tradeoff attacks.
-  | Argon2d -- ^ Argon2d is faster and uses data-depending memory access, which
-            -- makes it suitable for cryptocurrencies and applications with no
-            -- threats from side-channel timing attacks.
+  | Argon2d  -- ^ Argon2d is faster and uses data-depending memory access, which
+             -- makes it suitable for cryptocurrencies and applications with no
+             -- threats from side-channel timing attacks.
+  | Argon2id -- ^ Argon2id works as Argon2i for the first half of the first iteration
+             -- over the memory, and as Argon2d for the rest, thus providing both
+             -- side-channel attack protection and brute-force cost savings due to
+             -- time-memory tradeoffs.
+  deriving (Eq,Ord,Read,Show,Bounded,Generic,Typeable,Enum)
+
+-- | Version of the Argon2 algorithm.
+data Argon2Version
+  = Argon2Version10 -- ^ Version 1.0 (deprecated)
+  | Argon2Version13 -- ^ Version 1.3 (See [this announcment](https://www.ietf.org/mail-archive/web/cfrg/current/msg07948.html) for more details)
   deriving (Eq,Ord,Read,Show,Bounded,Generic,Typeable,Enum)
 
 -- | Parameters that can be adjusted to change the runtime performance of the
 -- hashing.
 data HashOptions =
-  HashOptions {hashIterations :: !Word32 -- ^ The time cost, which defines the amount of computation realized and therefore the execution time, given in number of iterations.
+  HashOptions { hashIterations  :: !Word32 -- ^ The time cost, which defines the amount of computation realized and therefore the execution time, given in number of iterations.
                                          --
                                          -- 'FFI.ARGON2_MIN_TIME' <= 'hashIterations' <= 'FFI.ARGON2_MAX_TIME'
-              ,hashMemory :: !Word32 -- ^ The memory cost, which defines the memory usage, given in kibibytes.
+              , hashMemory      :: !Word32 -- ^ The memory cost, which defines the memory usage, given in kibibytes.
                                      --
                                      -- max 'FFI.ARGON2_MIN_MEMORY' (8 * 'hashParallelism') <= 'hashMemory' <= 'FFI.ARGON2_MAX_MEMORY'
-              ,hashParallelism :: !Word32 -- ^ A parallelism degree, which defines the number of parallel threads.
+              , hashParallelism :: !Word32 -- ^ A parallelism degree, which defines the number of parallel threads.
                                           --
                                           -- 'FFI.ARGON2_MIN_LANES' <= 'hashParallelism' <= 'FFI.ARGON2_MAX_LANES' && 'FFI.ARGON_MIN_THREADS' <= 'hashParallelism' <= 'FFI.ARGON2_MAX_THREADS'
-              ,hashVariant :: !Argon2Variant -- ^ Which version of Argon2 to use.
+              , hashVariant     :: !Argon2Variant -- ^ Which variant of Argon2 to use.
+
+              , hashVersion     :: !Argon2Version -- ^ Which version of Argon2 to use for generating hashes.
+
+              , hashLength      :: !Word32        -- ^ Desired length of hash expressed in octets.
               }
   deriving (Eq,Ord,Read,Show,Bounded,Generic,Typeable)
 
@@ -76,27 +94,31 @@ data HashOptions =
 -- @
 -- 'defaultHashOptions' :: 'HashOptions'
 -- 'defaultHashOptions' =
---   'HashOptions' {'hashIterations' = 1
---               ,'hashMemory' = 2 ^ 17
---               ,'hashParallelism' = 4
---               ,'hashVariant' = 'Argon2i'}
+--   'HashOptions' { 'hashIterations' = 3
+--               , 'hashMemory' = 2 ^ 12 -- 4 MiB
+--               , 'hashParallelism' = 1
+--               , 'hashVariant' = 'Argon2i'
+--               , 'hashVersion' = 'Argon2V13'
+--               , 'hashLength'  = 2 ^ 5 -- 32 bytes
+--               }
 -- @
 defaultHashOptions :: HashOptions
-defaultHashOptions =
-  HashOptions {hashIterations = 1
-              ,hashMemory = 2 ^ 17
-              ,hashParallelism = 4
-              ,hashVariant = Argon2i}
+defaultHashOptions = HashOptions
+    { hashIterations  = 3
+    , hashMemory      = 2 ^ 12 -- 4 MiB
+    , hashParallelism = 1
+    , hashVariant     = Argon2i
+    , hashVersion     = Argon2Version13
+    , hashLength      = 32
+    }
 
 -- | Encode a password with a given salt and 'HashOptions' and produce a textual
 -- encoding of the result.
-hashEncoded :: HashOptions -- ^ Options pertaining to how expensive the hash is to calculate.
+hashEncoded :: HashOptions   -- ^ Options pertaining to how expensive the hash is to calculate.
             -> BS.ByteString -- ^ The password to hash. Must be less than 4294967295 bytes.
             -> BS.ByteString -- ^ The salt to use when hashing. Must be less than 4294967295 bytes.
-            -> T.Text -- ^ The encoded password hash.
-hashEncoded options password salt =
-  unsafePerformIO
-    (hashEncoded' options password salt FFI.argon2i_hash_encoded FFI.argon2d_hash_encoded)
+            -> TS.ShortText  -- ^ The encoded password hash.
+hashEncoded options password salt = unsafePerformIO $ hashEncoded' options password salt
 
 -- | Encode a password with a given salt and 'HashOptions' and produce a stream
 -- of bytes.
@@ -104,12 +126,12 @@ hash :: HashOptions -- ^ Options pertaining to how expensive the hash is to calc
      -> BS.ByteString -- ^ The password to hash. Must be less than 4294967295 bytes.
      -> BS.ByteString -- ^ The salt to use when hashing. Must be less than 4294967295 bytes.
      -> BS.ByteString -- ^ The un-encoded password hash.
-hash options password salt =
-  unsafePerformIO (hash' options password salt FFI.argon2i_hash_raw FFI.argon2d_hash_raw)
+hash options password salt = unsafePerformIO $ hash' options password salt
 
-variant :: a -> a -> Argon2Variant -> a
-variant a _ Argon2i = a
-variant _ b Argon2d = b
+variant :: a -> a -> a -> Argon2Variant -> a
+variant a _ _ Argon2i  = a
+variant _ b _ Argon2d  = b
+variant _ _ c Argon2id = c
 {-# INLINE variant #-}
 
 -- | Not all 'HashOptions' can necessarily be used to compute hashes. If you
@@ -127,7 +149,7 @@ data Argon2Exception
   | -- | Either too much or too little parallelism was requested via 'hasParallelism'.
     Argon2ParallelismOutOfRange !Word32 -- ^ The erroneous 'hashParallelism' value.
   | -- | An unexpected exception was throw. Please <https://github.com/ocharles/argon2/issues report this as a bug>!
-    Argon2Exception !Int32 -- ^ The =libargon2= error code.
+    Argon2Exception !CInt -- ^ The @libargon2@ error code.
   deriving (Typeable, Show)
 
 instance Exception Argon2Exception
@@ -137,71 +159,84 @@ type Argon2Encoded = Word32 -> Word32 -> Word32 -> CString -> CSize -> CString -
 hashEncoded' :: HashOptions
              -> BS.ByteString
              -> BS.ByteString
-             -> Argon2Encoded
-             -> Argon2Encoded
-             -> IO T.Text
-hashEncoded' options@HashOptions{..} password salt argon2i argon2d =
-  do let saltLen = fromIntegral (BS.length salt)
-         passwordLen = fromIntegral (BS.length password)
-     outLen <- fmap fromIntegral $ FFI.argon2_encodedlen
-                                              hashIterations
-                                              hashMemory
-                                              hashParallelism
-                                              saltLen
-                                              hashlen
-     out <- mallocBytes outLen
-     res <-
-       BS.useAsCString password $
-       \password' ->
-         BS.useAsCString salt $
-         \salt' ->
-           argon2 hashIterations
+             -> IO TS.ShortText
+hashEncoded' options@HashOptions{..} password salt = do
+    outLen <- FFI.argon2_encodedlen
+                  hashIterations
                   hashMemory
                   hashParallelism
-                  password'
-                  passwordLen
-                  salt'
-                  (fromIntegral saltLen)
-                  (fromIntegral hashlen)
-                  out
-                  (fromIntegral outLen)
-     handleSuccessCode res options password salt
-     fmap T.decodeUtf8 (BS.packCString out)
-  where argon2 = variant argon2i argon2d hashVariant
-        hashlen = 64
+                  saltLen
+                  hashLength
+                  argon2type
+
+    allocaBytes (fromIntegral outLen) $ \out -> do
+        res <- BS.useAsCString password $ \password' ->
+               BS.useAsCString salt $ \salt' ->
+               FFI.argon2_hash
+                   hashIterations
+                   hashMemory
+                   hashParallelism
+                   password'
+                   passwordLen
+                   salt'
+                   (fromIntegral saltLen)
+                   nullPtr
+                   (fromIntegral hashLength)
+                   out
+                   outLen
+                   argon2type
+                   argon2ver
+
+        handleSuccessCode res options password salt
+        res' <- TS.fromByteString <$> BS.packCString out
+        case res' of
+          Nothing -> fail "Argon2.hashEncoded: the impossible happened"
+          Just t  -> return t
+
+  where
+    saltLen     = fromIntegral (BS.length salt)
+    passwordLen = fromIntegral (BS.length password)
+    argon2type  = variant FFI.Argon2_i FFI.Argon2_d FFI.Argon2_id hashVariant
+    argon2ver   = case hashVersion of
+                    Argon2Version10 -> FFI.ARGON2_VERSION_10
+                    Argon2Version13 -> FFI.ARGON2_VERSION_13
 
 type Argon2Unencoded = Word32 -> Word32 -> Word32 -> CString -> CSize -> CString -> CSize -> CString -> CSize -> IO Int32
 
 hash' :: HashOptions
       -> BS.ByteString
       -> BS.ByteString
-      -> Argon2Unencoded
-      -> Argon2Unencoded
       -> IO BS.ByteString
-hash' options@HashOptions{..} password salt argon2i argon2d =
-  do let saltLen = fromIntegral (BS.length salt)
-         passwordLen = fromIntegral (BS.length password)
-         outLen = 512
-     out <- mallocBytes outLen
-     res <-
-       BS.useAsCString password $
-       \password' ->
-         BS.useAsCString salt $
-         \salt' ->
-           argon2 hashIterations
-                  hashMemory
-                  hashParallelism
-                  password'
-                  passwordLen
-                  salt'
-                  saltLen
-                  out
-                  (fromIntegral outLen)
-     handleSuccessCode res options password salt
-     BS.packCStringLen (out, outLen)
-  where argon2 = variant argon2i argon2d hashVariant
+hash' options@HashOptions{..} password salt =
+    allocaBytes (fromIntegral hashLength) $ \out -> do
+        res <- BS.useAsCString password $ \password' ->
+               BS.useAsCString salt $ \salt' ->
+               FFI.argon2_hash
+                   hashIterations
+                   hashMemory
+                   hashParallelism
+                   password'
+                   passwordLen
+                   salt'
+                   (fromIntegral saltLen)
+                   out
+                   (fromIntegral hashLength)
+                   nullPtr
+                   0
+                   argon2type
+                   argon2ver
 
-handleSuccessCode :: Int32
+        handleSuccessCode res options password salt
+        BS.packCStringLen (out, fromIntegral hashLength)
+  where
+    saltLen     = fromIntegral (BS.length salt)
+    passwordLen = fromIntegral (BS.length password)
+    argon2type  = variant FFI.Argon2_i FFI.Argon2_d FFI.Argon2_id hashVariant
+    argon2ver   = case hashVersion of
+                    Argon2Version10 -> FFI.ARGON2_VERSION_10
+                    Argon2Version13 -> FFI.ARGON2_VERSION_13
+
+handleSuccessCode :: CInt
                   -> HashOptions
                   -> BS.ByteString
                   -> BS.ByteString
@@ -220,29 +255,23 @@ handleSuccessCode res HashOptions{..} password salt =
            throwIO (Argon2IterationCountOutOfRange hashIterations)
          | a `elem` [FFI.ARGON2_MEMORY_TOO_LITTLE,FFI.ARGON2_MEMORY_TOO_MUCH] ->
            throwIO (Argon2MemoryUseOutOfRange hashMemory)
-         | a `elem`
-             [FFI.ARGON2_LANES_TOO_FEW
-             ,FFI.ARGON2_LANES_TOO_MANY
-             ,FFI.ARGON2_THREADS_TOO_FEW
-             ,FFI.ARGON2_THREADS_TOO_MANY] ->
+         | a `elem` [FFI.ARGON2_LANES_TOO_FEW,FFI.ARGON2_LANES_TOO_MANY
+                    ,FFI.ARGON2_THREADS_TOO_FEW,FFI.ARGON2_THREADS_TOO_MANY] ->
            throwIO (Argon2ParallelismOutOfRange hashParallelism)
          | otherwise -> throwIO (Argon2Exception a)
 
 -- | Verify that a given password could result in a given hash output.
 -- Automatically determines the correct 'HashOptions' based on the
 -- encoded hash (as produced by 'hashEncoded').
-verify
-  :: T.Text -> BS.ByteString -> Bool
-verify encoded password =
-  unsafePerformIO
-    (BS.useAsCString password $
-     \pwd ->
-       BS.useAsCString (T.encodeUtf8 encoded) $
-       \enc ->
-         do res <-
-              (variant FFI.argon2i_verify FFI.argon2d_verify v) enc
-                                                                pwd
-                                                                (fromIntegral (BS.length password))
-            return (res == FFI.ARGON2_OK))
-    where v | T.pack "$argon2i" `T.isPrefixOf` encoded = Argon2i
-            | otherwise = Argon2d
+verifyEncoded :: TS.ShortText -> BS.ByteString -> Bool
+verifyEncoded encoded password
+  -- c.f. https://github.com/P-H-C/phc-string-format/blob/master/phc-sf-spec.md
+  | "$argon2id$" `TS.isPrefixOf` encoded = unsafePerformIO $ go FFI.Argon2_id
+  | "$argon2i$"  `TS.isPrefixOf` encoded = unsafePerformIO $ go FFI.Argon2_i
+  | "$argon2d$"  `TS.isPrefixOf` encoded = unsafePerformIO $ go FFI.Argon2_d
+  | otherwise                                  = False
+  where
+    go v = BS.useAsCString password $ \pwd ->
+           BS.useAsCString (TS.toByteString encoded) $ \enc -> do
+               res <- FFI.argon2_verify enc pwd (fromIntegral (BS.length password)) v
+               return $! res == FFI.ARGON2_OK